Add input store type
